Begin implementing Bitboards
Hard to decide whether to implement bitbords as an abstracted class.

The overhead is definitely not worth it, but considering this is a
learning project, I think it makes sense to utilize C++ to it's full
capabilities.

Another option would be to essentially write C code but with whatever
syntactic sugar and goodies C++ has to offer, but if we're using an OOP
language, I think we really should **use** it.